flaskapp/requirements.txt
alembic==1.6.5
blueprint==3.4.2
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mysql==0.0.3
mysqlclient==2.0.3
PyMySQL==1.0.2
python-dateutil==2.8.1
python-decouple==3.4
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
WTForms==2.3.3
